+++ src/main/java/racingcar/model/Car.java
@@ -6,7 +6,7 @@
  * - 값이 3 이하일 경우 정지한다
  */
 
-package racingcar.domain;
+package racingcar.model;
 
 public class Car {
     private static final int MOVE_THRESHOLD = 4;
