+++ src/test/java/racingcar/CarTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.Car;
+import racingcar.model.Car;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
