+++ src/test/java/racingcar/RacingTest.java
@@ -3,8 +3,8 @@ package racingcar;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.Car;
-import racingcar.domain.Racing;
+import racingcar.model.Car;
+import racingcar.model.Racing;
 
 import java.util.ArrayList;
 import java.util.List;
